fix: resolve mypy type errors on Python 3.12 with types-requests installed
- news_sentiment.py: annotate params dict as dict[str, str] and convert
  pageSize to str so types-requests stub accepts the argument type
- backtest/__init__.py: wrap np.mean() results in float() to convert
  numpy floating[Any] to plain float; annotate recovery_idx as int to
  allow loop reassignment from int literal
- ai/lstm_model.py: wrap np.mean(raw_preds) in float() before passing
  to _unscale() which expects float, not numpy floating[Any]

+++ ai/lstm_model.py
@@ -188,7 +188,7 @@ class LSTMPipeline:
         x_last = self._prepare_input(values)
         raw_preds = self._raw_predictions(x_last)
 
-        pred_unscaled = self._unscale(np.mean(raw_preds), values)
+        pred_unscaled = self._unscale(float(np.mean(raw_preds)), values)
         return pred_unscaled - float(values[-1, self.target_idx])
 
     def prediction_uncertainty(
